+++ src/main/java/ua/lviv/bas/cinema/dao/SeatRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ua.lviv.bas.cinema.domain.Seat;
 
-public interface SeatRepository extends JpaRepository<Seat, Integer> {
+public interface SeatRepository extends JpaRepository<Seat, Long> {
 
 }
